backend/pyproject.toml
[tool.poetry]
name = "simple-app"
version = "0.1.0"
description = ""
authors = ["Konrad Zagozda <konrad.zagozda@grapeup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^1.1.0"
uvicorn = "^0.27.0"
psycopg = "^3.1.17"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pre-commit = "^3.6.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-django = {git = "https://github.com/rocioar/flake8-django.git"}
django-stubs = {version = "^4.2.7", extras = ["compatible-mypy"]}
pytest-django = "^4.7.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99
force_sort_within_sections = true

[tool.mypy]
strict = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
django_find_project = false
pythonpath = [".", "src"]